+++ frontend/components/TheFooter.vue
@@ -1,8 +1,36 @@
 <script setup lang="ts">
 import { useAuthStore } from '@/stores/AuthStore';
 
+// https://v8.i18n.nuxtjs.org/guide/per-component-translations
+const { t } = useI18n({
+  useScope: "local"
+});
+
 const authStore = useAuthStore();
 </script>
+
+<i18n lang="yaml">
+  en:
+    home: "Home"
+    add_bookmark: "Add Bookmark"
+    view_bookmarks: "View Bookmarks"
+    log_in: "Log In"
+    sign_up: "Sign Up"
+    user_profile: "User Profile"
+    manage_folders: "Manage Folders"
+    about_project: "About Project"
+    source_code: "Source Code"
+  ru:
+    home: "Главная"
+    add_bookmark: "Создать Закладку"
+    view_bookmarks: "Все Закладки"
+    log_in: "Войти"
+    sign_up: "Регистрация"
+    user_profile: "Профиль"
+    manage_folders: "Ваши Папки"
+    about_project: "О проекте"
+    source_code: "Исходный код"
+</i18n>
 
 <template>
   <footer class="footer">
@@ -11,6 +39,9 @@ const authStore = useAuthStore();
         <div class="column is-6">
           <p>
             &copy; <a href="https://github.com/hazadus/">hazadus</a>, 2023.
+          </p>
+          <p>
+            <LangSwitcher />
           </p>
         </div>
 
@@ -22,7 +53,9 @@ const authStore = useAuthStore();
                   <span class="icon">
                     <Icon name="material-symbols:roofing-outline" />
                   </span>
-                  <span>Home</span>
+                  <span>
+                    {{ t("home") }}
+                  </span>
                 </span>
               </NuxtLink>
             </li>
@@ -32,7 +65,9 @@ const authStore = useAuthStore();
                   <span class="icon">
                     <Icon name="material-symbols:bookmark-add" />
                   </span>
-                  <span>Add Bookmark</span>
+                  <span>
+                    {{ t("add_bookmark") }}
+                  </span>
                 </span>
               </NuxtLink>
             </li>
@@ -42,7 +77,9 @@ const authStore = useAuthStore();
                   <span class="icon">
                     <Icon name="material-symbols:bookmarks" />
                   </span>
-                  <span>View Bookmarks</span>
+                  <span>
+                    {{ t("view_bookmarks") }}
+                  </span>
                 </span>
               </NuxtLink>
             </li>
@@ -58,7 +95,9 @@ const authStore = useAuthStore();
                     <span class="icon">
                       <Icon name="material-symbols:login" />
                     </span>
-                    <span>Log In</span>
+                    <span>
+                      {{ t("log_in") }}
+                    </span>
                   </span>
                 </NuxtLink>
               </li>
@@ -69,7 +108,7 @@ const authStore = useAuthStore();
                       <Icon name="mdi:handshake-outline" />
                     </span>
                     <span>
-                      Sign Up
+                      {{ t("sign_up") }}
                     </span>
                   </span>
                 </NuxtLink>
@@ -82,7 +121,7 @@ const authStore = useAuthStore();
                     <Icon name="mdi:book-account" />
                   </span>
                   <span>
-                    User Profile
+                    {{ t("user_profile") }}
                   </span>
                 </span>
               </NuxtLink>
@@ -93,7 +132,9 @@ const authStore = useAuthStore();
                   <span class="icon">
                     <Icon name="material-symbols:folder-managed-rounded" />
                   </span>
-                  <span>Manage Folders</span>
+                  <span>
+                    {{ t("manage_folders") }}
+                  </span>
                 </span>
               </NuxtLink>
             </li>
@@ -109,7 +150,7 @@ const authStore = useAuthStore();
                     <Icon name="game-icons:gears" />
                   </span>
                   <span>
-                    About Project
+                    {{ t("about_project") }}
                   </span>
                 </span>
               </NuxtLink>
@@ -121,7 +162,7 @@ const authStore = useAuthStore();
                     <Icon name="mdi:github" />
                   </span>
                   <span>
-                    Source Code
+                    {{ t("source_code") }}
                   </span>
                 </span>
               </a>