+++ frontend/pages/folders.vue
@@ -2,6 +2,11 @@
 import { useAuthStore } from '@/stores/AuthStore';
 import type { Folder } from '@/types';
 
+// https://v8.i18n.nuxtjs.org/guide/per-component-translations
+const { t } = useI18n({
+  useScope: "local"
+});
+
 // "auth" middleware redirects user to login page if not authenticated:
 definePageMeta({
   middleware: "auth",
@@ -55,7 +60,7 @@ async function submitCreateFolderForm() {
 
   if (folderCreateError.value) {
     console.error("Error creating new folder: " + folderCreateError.value?.message);
-    alert("Something went wrong. Please try again!");
+    alert(t("error_something_went_wrong"));
     isLoading.value = false;
     return;
   }
@@ -69,10 +74,35 @@ async function submitCreateFolderForm() {
 fetchFolderData();
 </script>
 
-<template>
-  <Title>Manage folders | Bookmarks</Title>
+<i18n lang="yaml">
+  en:
+    page_title: "Manage folders"
+    page_header: "Manage folders"
+    error_header: "An error has occured while fetching folder data!"
+    error_something_went_wrong: "Something went wrong. Please try again!"
+    label_create_folder: "Create new folder"
+    input_placeholder: "New folder title"
+    info_header: "Information"
+    info_text: "If you delete a folder with bookmarks within, these bookmarks will be moved into Inbox folder."
+  ru:
+    page_title: "Управлять папками"
+    page_header: "Управлять папками"
+    error_header: "При получении данных о папках возникла ошибка!"
+    error_something_went_wrong: "Что-то пошло не так. Попробуйте ещё раз!"
+    label_create_folder: "Создать новую папку"
+    input_placeholder: "Название новой папки"
+    info_header: "Информация"
+    info_text: "Если вы удалите папку с закладками внутри, эти закладки будут перенесены в папку Входящие."
+</i18n>
 
-  <h2 class="title is-size-2">Manage folders</h2>
+<template>
+  <Title>
+    {{ t("page_title") }} | Bookmarks
+  </Title>
+
+  <h2 class="title is-size-2">
+    {{ t("page_header") }}
+  </h2>
 
   <div class="columns">
     <div class="column is-narrow">
@@ -83,7 +113,9 @@ fetchFolderData();
             <Icon name="material-symbols:dangerous-rounded" />
           </span>
           <span>
-            <strong>An error has occured while fetching folder data!</strong>
+            <strong>
+              {{ t("error_header") }}
+            </strong>
           </span>
         </div>
         <p class="block" v-for="error in errors" :key="error">
@@ -98,10 +130,12 @@ fetchFolderData();
 
       <form @submit.prevent="submitCreateFolderForm">
         <div class="field">
-          <label class="label">Create new folder:</label>
+          <label class="label">
+            {{ t("label_create_folder") }}:
+          </label>
           <div class="control">
-            <input class="input mr-1 new-folder-input" v-model="newFolderTitle" type="text" placeholder="New folder title"
-              maxlength="16" :disabled="isLoading">
+            <input class="input mr-1 new-folder-input" v-model="newFolderTitle" type="text"
+              :placeholder="t('input_placeholder')" maxlength="16" :disabled="isLoading">
             <button class="button is-success" :disabled="isLoading" :class="isLoading ? 'is-loading' : ''">
               <Icon name="mdi:plus" />
             </button>
@@ -116,10 +150,14 @@ fetchFolderData();
           <span class="icon has-text-info">
             <Icon name="mdi:information-variant-circle" />
           </span>
-          <span>Information</span>
+          <span>
+            <b>
+              {{ t("info_header") }}
+            </b>
+          </span>
         </div>
         <p class="block">
-          If you delete a folder with bookmarks within, these bookmarks will be moved into &laquo;Inbox&raquo; folder.
+          {{ t("info_text") }}
         </p>
       </BulmaNotification>
     </div>